+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.7.0] - 2026-05-20
 ### Changed
 - CycloneDX 1.7 Specification support
+- Added project custom field: safetyQualificationInput
 
 ## [1.6.0] - 2025-03-02
 ### Changed